update reference and usage docs, updated docstrings
