+++ docs/reference/testing/2-async-test-client.md
@@ -1,0 +1,19 @@
+
+# Async Test Client
+
+::: starlite.testing.client.async_client.AsyncTestClient
+    options:
+        members:
+            - __init__
+            - __aenter__
+            - request
+            - get
+            - post
+            - patch
+            - put
+            - delete
+            - options
+            - head
+            - set_session_data
+            - get_session_data
+            - portal
