+++ starlite/testing/client/base.py
@@ -120,36 +120,6 @@ class BaseTestClient(Generic[T]):
         return {cookie.key: cast("str", cookie.value) for cookie in backend._create_session_cookies(encoded_data)}
 
     async def _set_session_data_async(self, data: Dict[str, Any]) -> None:
-        """Set session data.
-
-        Args:
-            data: Session data
-
-        Returns:
-            None
-
-        Examples:
-            ```python
-            from starlite import Starlite, get
-            from starlite.middleware.session.memory_backend import MemoryBackendConfig
-
-            session_config = MemoryBackendConfig()
-
-
-            @get(path="/test")
-            def get_session_data(request: Request) -> Dict[str, Any]:
-                return request.session
-
-
-            app = Starlite(
-                route_handlers=[get_session_data], middleware=[session_config.middleware]
-            )
-
-            with AsyncTestClient(app=app, session_config=session_config) as client:
-                await client.set_session_data({"foo": "bar"})
-                assert await client.get("/test").json() == {"foo": "bar"}
-            ```
-        """
         mutable_headers = MutableScopeHeaders()
         await self.session_backend.store_in_message(
             scope_session=data,
@@ -166,33 +136,6 @@ class BaseTestClient(Generic[T]):
         self.cookies.update(cookies)  # type: ignore [union-attr]
 
     async def _get_session_data_async(self) -> Dict[str, Any]:
-        """Get session data.
-
-        Returns:
-            A dictionary containing session data.
-
-        Examples:
-            ```python
-            from starlite import Starlite, post
-            from starlite.middleware.session.memory_backend import MemoryBackendConfig
-
-            session_config = MemoryBackendConfig()
-
-
-            @post(path="/test")
-            def set_session_data(request: Request) -> None:
-                request.session["foo"] == "bar"
-
-
-            app = Starlite(
-                route_handlers=[set_session_data], middleware=[session_config.middleware]
-            )
-
-            with AsyncTestClient(app=app, session_config=session_config) as client:
-                await client.post("/test")
-                assert await client.get_session_data() == {"foo": "bar"}
-            ```
-        """
         return await self.session_backend.load_from_connection(
             connection=fake_asgi_connection(
                 app=self.app,
